+++ plugins/examples/discovery-node-role-extension/src/main/java/org/elasticsearch/example/discovery_node_role_extension/AdditionalRolePlugin.java
@@ -1,0 +1,56 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.elasticsearch.example.discovery_node_role_extension;
+
+import org.elasticsearch.cluster.node.DiscoveryNodeRole;
+import org.elasticsearch.common.settings.Setting;
+import org.elasticsearch.plugins.Plugin;
+
+import java.util.List;
+import java.util.Set;
+
+public class AdditionalRolePlugin extends Plugin {
+
+    static final Setting<Boolean> NODE_ADDITIONAL_SETTING = Setting.boolSetting("node.additional", false, Setting.Property.NodeScope);
+
+    @Override
+    public List<Setting<?>> getSettings() {
+        return List.of(NODE_ADDITIONAL_SETTING);
+    }
+
+    public static DiscoveryNodeRole ADDITIONAL_ROLE = new DiscoveryNodeRole("additional", "a") {
+
+        @Override
+        protected Setting<Boolean> roleSetting() {
+            return NODE_ADDITIONAL_SETTING;
+        }
+
+    };
+
+    public static class DiscoveryNodeRoleExtension implements org.elasticsearch.cluster.node.DiscoveryNodeRoleExtension {
+
+        @Override
+        public Set<DiscoveryNodeRole> roles() {
+            return Set.of(ADDITIONAL_ROLE);
+        }
+
+    }
+
+}
