+++ plugins/examples/discovery-node-role-extension/src/main/resources/META-INF/services/org.elasticsearch.cluster.node.DiscoveryNodeRoleExtension
@@ -1,0 +1,1 @@
+org.elasticsearch.example.discovery_node_role_extension.AdditionalRolePlugin$DiscoveryNodeRoleExtension
