+++ plugins/examples/discovery-node-role-extension/src/test/resources/rest-api-spec/test/discovery_node_role_extension/10_basic.yml
@@ -1,0 +1,15 @@
+"additional role":
+    - skip:
+        reason: "contains"
+        features: contains
+
+    - do:
+        cluster.state: {}
+
+    # get master node id
+    - set: { master_node: master }
+
+    - do:
+        nodes.info: {}
+
+    - match: { nodes.$master.roles.0: "additional" }
